Enabled operation buttons
The buttons with subclass of operation were given events
with a helper function. They now display a space, operation
symbol, and another space for enhanced visibility.
Moreover, the del button's function was modified to be
able to delete 3 spaces if it is detected that an operation
needs to be deleted.